devops-log-analysis-ai-agent/requirements.txt
langchain-google-genai
python-dotenv
langchain-core